refactor(vaccination-reminder): split create app logic
